Remove namespace from controllers
