+++ app/controllers/memberships_controller.rb
@@ -1,0 +1,8 @@
+class MembershipsController < ApplicationController
+  def destroy
+    membership = Membership.find(params[:id])
+    membership.destroy
+    notice :member_removed, person: membership.person, group: membership.group
+    redirect_to params[:referer] || '/'
+  end
+end
