+++ app/controllers/metrics/profiles_controller.rb
@@ -1,0 +1,15 @@
+module Metrics
+  class ProfilesController < ApplicationController
+    skip_before_action :ensure_user
+
+    def index
+      render json: {
+        'item' => [
+          {
+            'value' => Person.count
+          }
+        ]
+      }
+    end
+  end
+end
