+++ app/controllers/ping_controller.rb
@@ -1,0 +1,5 @@
+class PingController  < ActionController::Base
+  def index
+    render json: Deployment.info
+  end
+end
